pyproject.toml
[project]
name = "soc-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "scapy>=2.7.0",
    "requests>=2.32.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.26",
]
